improve webvh plugin
