+++ server/.env.example
@@ -1,3 +1,33 @@
 DOMAIN=''
 SECRET_KEY=''
 ENDORSER_MULTIKEY=''
+
+
+# POLICIES
+
+# Provide a known witness key
+KNOWN_WITNESS_KEY=null
+
+# Registry to used for known witnesses
+KNOWN_WITNESS_REGISTRY=https://registry.example.com
+
+# Require a known witness proof on attested resources
+WEBVH_ENDORSEMENT=true
+
+# Enforce a specific webvh version
+WEBVH_VERSION=1.0
+
+# Require at least 1 known witness signature
+WEBVH_WITNESS=true
+
+# URL for mandatory watcher service
+WEBVH_WATCHER=https://watcher.example.com
+
+# Boolean to enforce key rotation
+WEBVH_PREROTATION=true
+
+# Boolean to enforce portability
+WEBVH_PORTABILITY=true
+
+# Positive integer value representing the time range in hours acceptable for timestamps
+WEBVH_VALIDITY=1
