+++ server/app/plugins/__init__.py
@@ -1,5 +1,4 @@
 from .askar import AskarStorage, AskarVerifier
-from .didwebvh import DidWebVH
-from .policy import PolicyModule, PolicyError
+from .didwebvh import DidWebVH, PolicyError
 
-__all__ = ["AskarVerifier", "AskarStorage", "DidWebVH", "PolicyModule", "PolicyError"]
+__all__ = ["AskarVerifier", "AskarStorage", "DidWebVH", "PolicyError"]
