+++ server/app/plugins/didwebvh.py
@@ -2,28 +2,44 @@
 
 from fastapi import HTTPException
 from config import settings
+import requests
 from datetime import datetime
 from multiformats import multibase, multihash
 
 from app.utilities import digest_multibase
 import canonicaljson
 import json
+from asyncio import run as _await
 from did_webvh.core.state import DocumentState, verify_state_proofs
 
+class PolicyError(Exception):
+    """Policy error."""
+    pass
 
 class DidWebVH:
     """DID Web Verifiable History (DID WebVH) plugin."""
 
-    def __init__(self):
+    def __init__(self, active_policy=None, active_registry={}):
         """Initialize the DID WebVH plugin."""
-        self.prefix = settings.DID_WEBVH_PREFIX
-        self.method_version = f"{self.prefix}1.0"
-        self.did_string_base = self.prefix + r"{SCID}:" + settings.DOMAIN
+        self.prefix = 'did:webvh:'
+        self.scid_placeholder = settings.SCID_PLACEHOLDER
+        self.method_version = f"{self.prefix}{settings.WEBVH_VERSION}"
+        self.did_string_base = f'{self.prefix}{settings.SCID_PLACEHOLDER}:{settings.DOMAIN}'
+        self.active_policy = active_policy
+        self.known_witness_key = settings.KNOWN_WITNESS_KEY
+        self.known_witness_registry = active_registry
+        
+        # Reserved namespaces based on existing API routes
+        self.reserved_namespaces = ["policy"]
 
     def initial_document_state(self, params, document):
         """Return the latest document state."""
         doc_state = DocumentState.initial(params, document)
         return doc_state.history_line()
+
+    def placeholder_id(self, namespace, identifier):
+        """Return placeholder id."""
+        return f"did:webvh:{settings.SCID_PLACEHOLDER}:{settings.DOMAIN}:{namespace}:{identifier}"
 
     def get_document_state(self, log_entries, doc_state=None):
         """Return the latest document state."""
@@ -35,27 +51,18 @@ class DidWebVH:
         """Return the latest document state."""
         verify_state_proofs(state, prev_state)
 
-    def _init_state(self, did_doc):
-        return json.loads(json.dumps(did_doc).replace("did:web:", self.prefix + r"{SCID}:"))
-
-    def _generate_scid(self, log_entry):
+    def _generate_hash(self, scid_input):
         # https://identity.foundation/trustdidweb/#generate-scid
-        jcs = canonicaljson.encode_canonical_json(log_entry)
+        jcs = canonicaljson.encode_canonical_json(scid_input)
         multihashed = multihash.digest(jcs, "sha2-256")
         encoded = multibase.encode(multihashed, "base58btc")[1:]
         return encoded
 
-    def _generate_entry_hash(self, log_entry):
-        # https://identity.foundation/trustdidweb/#generate-entry-hash
-        jcs = canonicaljson.encode_canonical_json(log_entry)
-        multihashed = multihash.digest(jcs, "sha2-256")
-        encoded = multibase.encode(multihashed, "base58btc")[1:]
-        return encoded
-
-    def create_initial_did_doc(self, did_string):
-        """Create an initial DID document."""
-        did_doc = {"@context": [], "id": did_string}
-        return did_doc
+    def _find_witness_proof(self, proof_set, witness_id):
+        return [
+            proof for proof in proof_set 
+            if proof.get('verificationMethod').split('#')[0] == witness_id
+        ]
 
     def verify_resource(self, secured_resource):
         """Verify resource."""
@@ -123,3 +130,131 @@ class DidWebVH:
         identifier = did.split(":")[5]
         content_digest = resource_id.split("/")[-1]
         return f"{namespace}:{identifier}:{content_digest}"
+            
+    def namespace_available(self, namespace):
+        """Check if requested namespace is available."""
+        return False if namespace in self.reserved_namespaces else True
+
+    def load_known_witness_registry(self, registry):
+        """Load known witness registry."""
+        
+        self.known_witness_registry = registry
+        
+        if self.known_witness_key:
+            witness_id = f'did:key:{self.known_witness_key}'
+            if witness_id not in self.known_witness_registry:
+                self.known_witness_registry[witness_id] = {
+                    'name': 'Default Server Witness'
+                }
+
+    def cache_known_witness_registry(self):
+        """Cache known witness registry."""
+        if self.known_witness_registry_url:
+            r = requests.get(self.known_witness_registry_url)
+            registry = r.json()
+        
+        for witness in registry.get('registry'):
+            if not witness.startswith('did:key:'):
+                raise PolicyError(f"Invalid witness registry: {self.known_witness_registry_url}")
+            
+        return self.known_witness_registry | registry.get('registry')
+
+
+    def validate_known_witness(self, document_state, witness_signature):
+        """Validate known witness."""
+        witness_id = document_state.params.get('witness').get('witnesses').get(0)
+            
+        if not witness_id:
+            raise PolicyError("No witness")
+        
+        if witness_id not in self.known_witness_registry:
+            self.cache_known_witness_registry()
+            if witness_id not in self.known_witness_registry:
+                raise PolicyError("Unknown witness")
+        
+        witness_proof = self._find_witness_proof(
+            witness_signature.get('signatures'), 
+            witness_id
+        )
+            
+        if not witness_proof:
+            raise PolicyError("No witness proof")
+        
+
+    def create_did(self, log_entry, witness_signature=None):
+        """Apply policies to DID creation."""
+
+        document_state = self.get_document_state([log_entry])
+        self.verify_state_proofs(document_state)
+        
+        witness_rules = document_state.witness_rule
+        # if self.webvh_witness:
+        #     self.validate_known_witness(document_state, witness_signature)
+            
+        log_entries = [document_state.history_line()]
+        witness_file = [witness_signature]
+            
+        return log_entries, witness_file
+
+    def update_did(self, log_entry, log_entries, witness_signature=None, prev_witness_file=None):
+        """Apply policies to DID updates."""
+        prev_document_state = self.get_document_state(log_entries)
+        if prev_document_state.params.get("deactivated"):
+            raise PolicyError("DID is deactivated")
+        
+        document_state = self.get_document_state([log_entry], prev_document_state)
+        self.verify_state_proofs(document_state, prev_document_state)
+        
+        if prev_document_state.next_key_hashes:
+            document_state._validate_key_rotation(
+                prev_document_state.next_key_hashes, document_state.update_keys
+            )
+        
+        witness_rules = prev_document_state.witness_rule
+        # if self.webvh_witness:
+        #     self.validate_known_witness(document_state, witness_signature)
+            
+        if document_state.deactivated:
+            self.deactivate_did()
+            
+        log_entries.append(document_state.history_line())
+        # witness_file = [prev_witness_file | witness_signature]
+        witness_file = [witness_signature]
+            
+        return log_entries, witness_file
+
+    def deactivate_did(self, log_entry, witness_signature=None):
+        """Apply policies to DID deactivation."""
+        return
+
+    def proof_options(self):
+        """Create new proof options."""
+        return {
+            'type': 'DataIntegrityProof',
+            'cryptosuite': 'eddsa-jcs-2022',
+            'proofPurpose': 'assertionMethod'
+        }
+
+    def parameters(self):
+        """Create policy driven parameters."""
+        
+        server_parameter = {
+            'scid': settings.SCID_PLACEHOLDER,
+            'method': f'did:webvh:{settings.WEBVH_VERSION}',
+            'portability': self.active_policy.get('portability'),
+            'updateKeys': []
+        }
+        
+        if self.active_policy.get('prerotation'):
+            server_parameter['nextKeyHashes'] = []
+            
+        if self.active_policy.get('witness'):
+            server_parameter['witness'] = {
+                'threshold': 1,
+                'witnesses': [{'id': witness} for witness in self.known_witness_registry]
+            }
+            
+        if self.active_policy.get('watcher'):
+            server_parameter['watchers'] = [self.active_policy.get('watcher')]
+            
+        return server_parameter
